+++ model/nist-lte-ue-cphy-sap.h
@@ -107,7 +107,7 @@ public:
   /**
    * \param dlBandwidth the DL bandwidth in number of PRBs
    */
-  virtual void SetDlBandwidth (uint8_t dlBandwidth) = 0;
+  virtual void SetDlBandwidth (uint16_t dlBandwidth) = 0;
 
   /** 
    * \brief Configure uplink (normally done after reception of SIB2)
@@ -115,7 +115,7 @@ public:
    * \param ulEarfcn the uplink carrier frequency (EARFCN)
    * \param ulBandwidth the UL bandwidth in number of PRBs
    */
-  virtual void ConfigureUplink (uint16_t ulEarfcn, uint8_t ulBandwidth) = 0;
+  virtual void ConfigureUplink (uint16_t ulEarfcn, uint16_t ulBandwidth) = 0;
 
   /**
    * \brief Configure referenceSignalPower
@@ -365,8 +365,8 @@ public:
   virtual void StartCellSearch (uint16_t dlEarfcn);
   virtual void SynchronizeWithEnb (uint16_t cellId);
   virtual void SynchronizeWithEnb (uint16_t cellId, uint16_t dlEarfcn);
-  virtual void SetDlBandwidth (uint8_t dlBandwidth);
-  virtual void ConfigureUplink (uint16_t ulEarfcn, uint8_t ulBandwidth);
+  virtual void SetDlBandwidth (uint16_t dlBandwidth);
+  virtual void ConfigureUplink (uint16_t ulEarfcn, uint16_t ulBandwidth);
   virtual void ConfigureReferenceSignalPower (int8_t referenceSignalPower);
   virtual void SetRnti (uint16_t rnti);
   virtual void SetTransmissionMode (uint8_t txMode);
@@ -436,14 +436,14 @@ NistMemberLteUeCphySapProvider<C>::SynchronizeWithEnb (uint16_t cellId, uint16_t
 
 template <class C>
 void
-NistMemberLteUeCphySapProvider<C>::SetDlBandwidth (uint8_t dlBandwidth)
+NistMemberLteUeCphySapProvider<C>::SetDlBandwidth (uint16_t dlBandwidth)
 {
   m_owner->DoSetDlBandwidth (dlBandwidth);
 }
 
 template <class C>
 void 
-NistMemberLteUeCphySapProvider<C>::ConfigureUplink (uint16_t ulEarfcn, uint8_t ulBandwidth)
+NistMemberLteUeCphySapProvider<C>::ConfigureUplink (uint16_t ulEarfcn, uint16_t ulBandwidth)
 {
   m_owner->DoConfigureUplink (ulEarfcn, ulBandwidth);
 }
